+++ github_api.sh
@@ -10,7 +10,7 @@ github_api() {
   _shlib_header=""
   case "$_shlib_source_url" in
     https://api.github.com*)
-      test -z "$GITHUB_TOKEN" || _shlib_header="Authorization: token $GITHUB_TOKEN"
+      test -z "${GITHUB_TOKEN-}" || _shlib_header="Authorization: token $GITHUB_TOKEN"
       ;;
   esac
   http_download "$_shlib_local_file" "$_shlib_source_url" "$_shlib_header"